# backend/src/main/java/be/icc/metamind/user/ProfileController.java
package be.icc.metamind.user;

import jakarta.validation.Valid;

import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PutMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/api/v1/users")
public class ProfileController {
	private final AccountService service;

	public ProfileController(AccountService service) {
		this.service = service;
	}

	@GetMapping("/{id}/profile")
	public UserResponse getProfile(@PathVariable long id, @RequestHeader("Authorization") String authorization) {
		service.authenticateSelfOrAdmin(id, authorization);
		return service.getProfile(id);
	}

	@PutMapping("/{id}/profile")
	public UserResponse updateProfile(@PathVariable long id, @RequestHeader("Authorization") String authorization, @Valid @RequestBody UpdateProfileRequest request) {
		service.authenticateSelfOrAdmin(id, authorization);
		return service.updateProfile(id, request);
	}

	@DeleteMapping("/{id}")
	public UserResponse requestDeletion(@PathVariable long id, @RequestHeader("Authorization") String authorization) {
		service.authenticateSelfOrAdmin(id, authorization);
		return service.requestAccountDeletion(id);
	}
}